Refactor plants module and add API integration
- Moved plants module to new API directory structure
- Created PlantsController to handle plant data retrieval
- Implemented PlantsService with Trefle API integration
- Added fetch method with caching for plant data
- Updated import paths in app module
