+++ src/app/api/plants/plants.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
-import { PlantsController } from '@/modules/plants/plants.controller';
-import { PlantsService } from '@/modules/plants/plants.service';
+import { PlantsController } from '@/app/api/plants/plants.controller';
+import { PlantsService } from '@/app/api/plants/plants.service';
 
 @Module({
   imports: [HttpModule],
