+++ src/app/api/plants/plants.service.ts
